// src/app/shared/generated/model/getTaskResponse.ts
/**
 * onecx-human-task-bff internal service
 * Backend-For-Frontend (BFF) API for onecx-human-task. Provides endpoints for the Angular frontend to search tasks, view task details, and perform accept/decline actions.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Task } from './task';


export interface GetTaskResponse { 
    resource: Task;
}