// src/app/shared/generated/model/task.ts
/**
 * onecx-human-task-bff internal service
 * Backend-For-Frontend (BFF) API for onecx-human-task. Provides endpoints for the Angular frontend to search tasks, view task details, and perform accept/decline actions.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProviderType } from './providerType';
import { TaskStatus } from './taskStatus';


export interface Task { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    title?: string;
    description?: string;
    status?: TaskStatus;
    providerType?: ProviderType;
    /**
     * The task ID in the external provider system
     */
    providerTaskId?: string;
    /**
     * URL of the specific provider instance
     */
    providerURL?: string;
    /**
     * Arbitrary key-value pairs provided by users during accept/decline
     */
    customInput?: { [key: string]: string; };
}